# Make 1D integer sorting work in parallel