@DamianSzwichtenberg 2023-04-27
> You are also updating fbgemm submodule, usually that's done in a separate PR

@ngimel I was intending to do this with ghstack, but I don't have write access. Anyway, I can still separate those commits, just let me know if I should do so.