@DamianSzwichtenberg 2023-04-28
@ngimel 
> Yes, please do it in separate PRs, we can land one after another

Done via #100236

> The linker errors in the CI are real

Yeah, I'm aware, I struggled to reproduce those errors (everything works fine on my environment). Is it possible for you to share with me `pytorch/pytorch-linux-focal-py3.8-gcc7` docker image?