@malfet 2023-04-28
@DamianSzwichtenberg those images are available on ghcr.io, for example one your refer to is available at https://github.com/orgs/pytorch/packages/container/ci-image/88094155?tag=pytorch-linux-focal-py3.8-gcc7-78844eeb8b002d24db840866dc128c3fe7ad96f8