@DamianSzwichtenberg 2023-05-05
> It should be OK to have this file removed: `aten/src/ATen/native/cpu/radix_sort.h` This is also one of my original initiative :)

I'm planning to clean-up scatter-related code in a separate PR. 😉