@mingfeima 2023-10-23
@DamianSzwichtenberg, Meta engineers found an heap overflow from https://github.com/pytorch/FBGEMM/pull/1672 which results into https://github.com/pytorch/pytorch/issues/111189 and https://github.com/pytorch/pytorch/pull/111099. I believe a fix has been merged https://github.com/pytorch/FBGEMM/pull/2075. Could you please double check and make sure that corresponding test cases are also updated to reflect this error in torch.sort().